+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/controller/model/FilterDefinition.kt
@@ -1,7 +1,6 @@
 package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model
 
 import com.google.gson.annotations.SerializedName
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.DynamicFilterOption
 
 data class FilterDefinition(
   val type: FilterType,
@@ -13,4 +12,5 @@ data class FilterDefinition(
   val defaultValue: String? = null,
   val min: String? = null,
   val max: String? = null,
+  val interactive: Boolean? = false,
 )