+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/service/ReportDefinitionSummaryMapper.kt
@@ -1,7 +1,7 @@
 package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service
 
 import org.springframework.stereotype.Component
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.DashboardDefinition
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.DashboardDefinitionSummary
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.RenderMethod
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.ReportDefinitionSummary
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.VariantDefinitionSummary
@@ -10,7 +10,7 @@ import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.Product
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.Report
 
 @Component
-class ReportDefinitionSummaryMapper(val dashboardDefinitionService: DashboardDefinitionService) {
+class ReportDefinitionSummaryMapper {
 
   fun map(
     productDefinition: ProductDefinition,
@@ -22,11 +22,8 @@ class ReportDefinitionSummaryMapper(val dashboardDefinitionService: DashboardDef
     variants = productDefinition.report
       .filter { renderMethod == null || it.render.toString() == renderMethod.toString() }
       .map { map(it) },
-    dashboards = map(productDefinition.dashboards),
+    dashboards = productDefinition.dashboards?.map { map(it) },
   )
-
-  private fun map(dashboards: List<Dashboard>?): List<DashboardDefinition>? =
-    dashboards?.map { dashboardDefinitionService.toDashboardDefinition(it) }
 
   private fun map(
     report: Report,
@@ -35,6 +32,16 @@ class ReportDefinitionSummaryMapper(val dashboardDefinitionService: DashboardDef
       id = report.id,
       name = report.name,
       description = report.description,
+    )
+  }
+
+  private fun map(
+    dashboard: Dashboard,
+  ): DashboardDefinitionSummary {
+    return DashboardDefinitionSummary(
+      id = dashboard.id,
+      name = dashboard.name,
+      description = dashboard.description,
     )
   }
 }