+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 Below you can find the changes included in each release.
 
+## 7.1.1
+Fix issue with dashboard query execution.
+
+## 7.1.0
+Added filters to getStatementResults endpoint and dashboard definition.
+
 ## 7.0.5
 Added endpoint to cancel running dashboard queries.